requirements.txt
Django==1.6.4
djangorestframework==2.3.13
pytz==2014.2
